Consolidate tests into one folder and add setup
Moves all tests into one folder and adds the necessary setup to run them
