+++ .github/workflows/ci.yml
@@ -23,9 +23,6 @@ jobs:
       - name: Run Unit Tests
         run:  make test
 
-      - name: Run Integration Tests
-        run: echo TODO-TODO-TODO-TODO-TODO-TODO-TODO-TODO-TODO-TODO-TODO-TODO
-
       - name: Configure AWS credentials
         uses: aws-actions/configure-aws-credentials@v1-node16
         with:
