+++ docker-compose.dev.yml
@@ -12,3 +12,10 @@ services:
       timeout: 3s
       retries: 30
       start_period: 20s
+  navigator-search-indexer:
+    build:
+      context: .
+    entrypoint: []
+    environment:
+      VESPA_INSTANCE_URL: "http://vespaindexertest:8080/"
+      DEVELOPMENT_MODE: "True"
