+++ tests/cli/test_data/index_data_input/test_html.npy
