+++ tests/cli/test_data/index_data_input/test_no_content_type.json
