+++ tests/cli/test_index_data.py
