+++ tests/test_integration.py
@@ -1,12 +1,13 @@
 import os
 
 from click.testing import CliRunner
+
 from cli.index_data import run_as_cli
 
 
 def test_integration():
     fixture_dir = os.path.join(os.path.dirname(os.path.abspath(__file__)), "s3_fixtures")
-
+    
     runner = CliRunner()
     result = runner.invoke(
         run_as_cli,
@@ -15,10 +16,6 @@ def test_integration():
             "--index-type",
             "vespa",
         ],
-        env={
-            "VESPA_INSTANCE_URL": "http://localhost:8080/",
-            "DEVELOPMENT_MODE": "True"
-        },
     )
 
     assert result.exit_code == 0, (
